// src/jobs/job.interface.ts
export interface JobData {
  link: string;
  job_type: string;
  title: string;
  company: string;
  salary: string;
  location: string;
  experience: string;
  level: string;
  workingForm: string;
  deadline: string;
  shift: string;
  degree: string;
  age: string;
  quantity: string;
  field: string;
  description: string;
  requirement: string;
  benefit: string;
  companyLogo: string;
  companySize: string;
  companyHeadquarters: string;
  contactName: string;
  contactPhone: string;
}